refactored shared code for the services and invoke Local Lambda on request
Details:
* Moved shared code between APIGW and Lambda service to a Base Class
* Updated the Invoke function on the LocalLambdaService to invoke the Local Lambda based on the request
* Updated Functional tests for invoking the Local Lambda
